Search for right DOM node on override
The override card received the DOM node to manipulate when the card
was set.  The DOM node could disappear by that time.  We've
reimplemented the logic to detect the node to override.  This is far
from a clean solution and it should be revisited in the API provided
to this component.